fix: VideoProcessor 整数除法导致进度始终显示 0%
- frameIndex / frameCount 两个 int 操作数除法结果为 0 直至最后一帧
- 改为 (double)frameIndex / frameCount 实现平滑进度

+++ SekaiToolsCore/VideoProcessor.cs
@@ -220,7 +220,7 @@ public class VideoProcessor
                 }
 
                 frameIndex = (int)capture.Get(CapProp.PosFrames);
-                var progress = frameCount > 0 ? frameIndex / frameCount : 0;
+                var progress = frameCount > 0 ? (double)frameIndex / frameCount : 0;
 
                 // 节流进度回调（200ms）
                 EmitProgressIfNeeded(progress);
